+++ ReleaseNotes.md
@@ -1,5 +1,10 @@
 # Release notes of Ampersand
 
+## v3.11.1 (3 august 2018)
+
+* [Issue #814](https://github.com/AmpersandTarski/Ampersand/issues/814) More clear error message.
+* [Issue #798](https://github.com/AmpersandTarski/Ampersand/issues/798) Refactor code, to remove some double code.
+
 ## v3.11.0 (6 july 2018)
 
 * Upgrade to LTS-11.13